requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0

djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
gunicorn==20.0.4